requirements.txt
requests==2.31.0
pydantic==2.5.0
duckdb==1.1.3
apache-airflow==2.7.2
python-dotenv==1.0.0
pytest==7.4.3
openai==2.15.0
pinecone
streamlit
